Code refactor with ConnectWPComAccountCard.
