+++ js/src/setup-mc/setup-stepper/setup-accounts/index.js
@@ -20,7 +20,7 @@ import WordPressDotComAccount from './wordpressdotcom-account';
 import GoogleAccountCard from '.~/components/google-account-card';
 import GoogleMCAccount from './google-mc-account';
 import Faqs from './faqs';
-import WPComAccountCard from './wordpressdotcom-account/wpcom-account-card';
+import ConnectWPComAccountCard from './wordpressdotcom-account/connect-wpcom-account-card';
 
 const SetupAccounts = ( props ) => {
 	const { onContinue = () => {} } = props;
@@ -64,7 +64,7 @@ const SetupAccounts = ( props ) => {
 			>
 				<VerticalGapLayout size="large">
 					<WordPressDotComAccount />
-					<WPComAccountCard />
+					<ConnectWPComAccountCard />
 					<GoogleAccountCard disabled={ isGoogleAccountDisabled } />
 				</VerticalGapLayout>
 			</Section>
