+++ js/src/setup-mc/setup-stepper/setup-accounts/wordpressdotcom-account/wpcom-account-card/index.js
@@ -7,32 +7,9 @@ import { __ } from '@wordpress/i18n';
  * Internal dependencies
  */
 import AccountCard from '.~/components/account-card';
-import AppButton from '.~/components/app-button';
-import useDispatchCoreNotices from '.~/hooks/useDispatchCoreNotices';
-import useApiFetchCallback from '.~/hooks/useApiFetchCallback';
 import wpLogoURL from './wp-logo.svg';
 
-const WPComAccountCard = () => {
-	const { createNotice } = useDispatchCoreNotices();
-	const [ fetchJetpackConnect, { loading, data } ] = useApiFetchCallback( {
-		path: '/wc/gla/jetpack/connect',
-	} );
-
-	const handleConnectClick = async () => {
-		try {
-			const d = await fetchJetpackConnect();
-			window.location.href = d.url;
-		} catch ( error ) {
-			createNotice(
-				'error',
-				__(
-					'Unable to connect your WordPress.com account. Please try again later.',
-					'google-listings-and-ads'
-				)
-			);
-		}
-	};
-
+const WPComAccountCard = ( props ) => {
 	return (
 		<AccountCard
 			appearance={ {
@@ -49,20 +26,7 @@ const WPComAccountCard = () => {
 				),
 				title: 'WordPress.com',
 			} }
-			description={ __(
-				'Required to connect with Google',
-				'google-listings-and-ads'
-			) }
-			indicator={
-				<AppButton
-					isSecondary
-					loading={ loading || data }
-					eventName="gla_wordpress_account_connect_button_click"
-					onClick={ handleConnectClick }
-				>
-					{ __( 'Connect', 'google-listings-and-ads' ) }
-				</AppButton>
-			}
+			{ ...props }
 		/>
 	);
 };
